requirements.txt
requests==2.31.0
pandas==2.0.3
numpy==1.24.4
psycopg2-binary==2.9.9
